Single provider and dependencies added.
Update as on 02/12/2022.
Limitaion #1 removed. Now the providers are not created in each module. The providers are created in the root folder and then the required providers are passed to the corresponding modules.
Since the provider is called from the root folder, dependencies on modules can be added as well.

Limitations as on 02-Dec-2022:
1. The credentials are currently stored in variable.tf file. This needs to be hidden so that the credentials are never exposed to anyone.
2. This code currently has the basic items required in a Snowflake environment. Extra features need to be added and implemented successfully.